library/requirements.txt
requests
kubernetes-py==0.26
